Revise general architecture
It was necessary to revise the file hierarchy.
The app itself is running and circular imports
are avoided due to an init script setting up
the database before the models are used (have
an import of db)
Routes are also extracted to a separat file.

TODO:
Altough the database connection could be
established, there is still a problem with
the migration and set up of tables.