Add functionality to start evaluation
A cookie with the user id is set to track
the user and match the courses and store
the choices.
A functionality to go from one evaluation
to the next needs to be added.
Additionally, the number of courses already
evaluated needs to be stored or checked.
Every user should evaluate 10 courses.

TODO:
- add functionalities and move to next
evaluation website
- add functionality to count how many courses
are evaluated by the current user!
- improve styling of landing page
- add imprint at the bottom