Select approx. same number of keywords
We identified a possible bias if the numbers
of keywords in the two lists differ too much.
Users tend to choose the list according to
the number of keywords in the list in some
prior UX test.
Now between 4 and 6 keywords are selected
randomly from all keywords independently
if it is the list of manual assigned or
generated keywords. If the list has less
than 6 keywords the list is displayed as is.